[FEAT]: 장소 투표 호스트 확정 기능 구현
호스트/게스트 버튼 분기를 위해 시트 콘텐츠를
PlaceVoteSheetContent로 분리하고 Countdown 공통 모델로 전환
