+++ Projects/Data/API/Sources/VoteEndpoint.swift
@@ -17,6 +17,7 @@ public enum VoteEndpoint: EndPoint {
     case confirmDateVote(meetingId: Int, ConfirmDateVoteRequestDTO)
     case startPlaceVote(meetingId: Int, StartPlaceVoteRequestDTO)
     case submitPlaceVote(meetingId: Int, SubmitPlaceVoteRequestDTO)
+    case confirmPlaceVote(meetingId: Int)
     case fetchPlaceVoteTravelBurden(meetingId: Int, placeId: Int)
     case fetchPlaceVoteParticipants(meetingId: Int)
 
@@ -32,6 +33,7 @@ public enum VoteEndpoint: EndPoint {
         case .confirmDateVote(let meetingId, _): return "/api/v1/meetings/\(meetingId)/date-vote/confirm"
         case .startPlaceVote(let meetingId, _): return "/api/v1/meetings/\(meetingId)/place-vote"
         case .submitPlaceVote(let meetingId, _): return "/api/v1/meetings/\(meetingId)/place-vote/submit"
+        case .confirmPlaceVote(let meetingId): return "/api/v1/meetings/\(meetingId)/place-confirm"
         case .fetchPlaceVoteTravelBurden(let meetingId, let placeId):
             return "/api/v1/meetings/\(meetingId)/place-vote/\(placeId)/travel-burden"
         case .fetchPlaceVoteParticipants(let meetingId):
@@ -49,6 +51,7 @@ public enum VoteEndpoint: EndPoint {
         case .confirmDateVote: return .patch
         case .startPlaceVote: return .post
         case .submitPlaceVote: return .post
+        case .confirmPlaceVote: return .post
         case .fetchPlaceVoteTravelBurden: return .get
         case .fetchPlaceVoteParticipants: return .get
         }
@@ -64,6 +67,7 @@ public enum VoteEndpoint: EndPoint {
         case .confirmDateVote(_, let dto): return .requestJSONEncodable(body: dto)
         case .startPlaceVote(_, let dto): return .requestJSONEncodable(body: dto)
         case .submitPlaceVote(_, let dto): return .requestJSONEncodable(body: dto)
+        case .confirmPlaceVote: return .requestPlain
         case .fetchPlaceVoteTravelBurden: return .requestPlain
         case .fetchPlaceVoteParticipants: return .requestPlain
         }
