+++ Projects/Domain/DataInterface/Sources/VoteRepositoryProtocol.swift
@@ -14,6 +14,7 @@ public protocol VoteRepositoryProtocol: Sendable {
     func confirmDateVote(meetingId: Int, optionId: Int) async throws
     func startPlaceVote(meetingId: Int, durationDays: Int) async throws
     func submitPlaceVote(meetingId: Int, placeIds: [Int]) async throws
+    func confirmPlaceVote(meetingId: Int) async throws
     func fetchPlaceVoteTravelBurden(meetingId: Int, placeId: Int) async throws -> PlaceVoteTravelBurden
     func fetchPlaceVoteParticipants(meetingId: Int) async throws -> [PlaceVoteParticipant]
 }
