+++ Projects/Domain/UseCase/Sources/Vote/ConfirmPlaceVoteUseCase.swift
@@ -1,0 +1,8 @@
+//
+//  ConfirmPlaceVoteUseCase.swift
+//  UseCase
+//
+
+public protocol ConfirmPlaceVoteUseCase: Sendable {
+    func execute(meetingId: Int) async throws
+}
