+++ Projects/Presentation/HomeFeature/Sources/GroupDetail/PlaceVoteParticipation/PlaceVoteParticipationView.swift
@@ -11,7 +11,6 @@ import ComposableArchitecture
 import CoreDependencies
 import DesignSystem
 import Entity
-import Utill
 
 /// 약속 장소 투표 참여 화면.
 /// 전체 화면 카카오맵 위에 NavigationPage 와 MapBottomSheet 가 겹쳐진 구조다.
@@ -66,7 +65,7 @@ struct PlaceVoteParticipationView: View {
             }
             .onVisibleHeightChanged { sheetCoveredHeight = $0 }
 
-            PlaceVoteButtonArea(store: store, onComplete: dismissCover)
+            PlaceVoteButtonArea(store: store)
                 .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .bottom)
                 .ignoresSafeArea(edges: .bottom)
         }
@@ -204,82 +203,13 @@ struct PlaceVoteParticipationView: View {
     }
 }
 
-// MARK: - Sheet Content
-
-private struct PlaceVoteSheetContent: View {
-    let store: StoreOf<PlaceVoteParticipationFeature>
-    let onFocusPlace: (MapCoordinate) -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: Spacing.spacing300) {
-            PlaceVoteHeader(store: store)
-
-            VStack(spacing: Spacing.spacing200) {
-                ForEach(store.candidates) { candidate in
-                    PlaceVoteRow(
-                        candidate: candidate,
-                        mode: store.mode,
-                        isSelected: store.selectedPlaceIds.contains(candidate.id),
-                        isTop: store.topPlaceId == candidate.id,
-                        onTap: { focus(candidate) },
-                        onSelect: { store.send(.placeSelected(candidate.id)) }
-                    )
-                }
-            }
-        }
-        .padding(.top, Spacing.spacing100)
-        .padding(.bottom, UIScreen.safeAreaBottom + PlaceVoteButtonArea.height)
-    }
-
-    /// 멤버별 경로를 조회하고, 후보 좌표가 있으면 지도 포커싱을 요청한다.
-    private func focus(_ candidate: PlaceVoteCandidate) {
-        store.send(.placeRowTapped(candidate.id))
-
-        guard
-            let latitude = candidate.latitude,
-            let longitude = candidate.longitude
-        else { return }
-
-        onFocusPlace(MapCoordinate(latitude: latitude, longitude: longitude))
-    }
-}
-
-// MARK: - Sheet Header
-
-private struct PlaceVoteHeader: View {
-    let store: StoreOf<PlaceVoteParticipationFeature>
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: Spacing.spacing150) {
-            BangawoText("약속 장소 투표하기", textStyle: .titleLarge)
-                .foregroundStyle(Colors.gray800)
-
-            HStack(spacing: 0) {
-                DeadlineDescription(deadline: store.deadline)
-
-                if store.mode == .voted {
-                    Button {
-                        store.send(.participantsButtonTapped)
-                    } label: {
-                        BangawoText(" · 현재 \(store.votedCount)명 참여", textStyle: .bodySmall)
-                            .foregroundStyle(Colors.gray700)
-                    }
-                    .buttonStyle(.plain)
-                }
-            }
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
-
 // MARK: - Button Area
 
-private struct PlaceVoteButtonArea: View {
+struct PlaceVoteButtonArea: View {
     /// 버튼(large minHeight) + 하단 패딩. safeArea 는 별도로 더한다.
     static let height: CGFloat = Spacing.spacing800 + Spacing.spacing200
 
     let store: StoreOf<PlaceVoteParticipationFeature>
-    let onComplete: () -> Void
 
     var body: some View {
         Group {
@@ -296,18 +226,27 @@ private struct PlaceVoteButtonArea: View {
                 }
 
             case .voted:
-                ActionButton(
-                    buttonLayout: .dual(
-                        primaryTitle: "완료",
-                        primaryAction: {
-                            store.send(.completeButtonTapped)
-                            onComplete()
-                        },
-                        secondaryTitle: "다시 투표하기",
-                        secondaryAction: { store.send(.revoteButtonTapped) },
-                        arrangement: .horizontal
+                if store.isHost {
+                    ActionButton(
+                        buttonLayout: .dual(
+                            primaryTitle: "확정하기",
+                            primaryAction: { store.send(.confirmButtonTapped) },
+                            secondaryTitle: "다시 투표하기",
+                            secondaryAction: { store.send(.revoteButtonTapped) },
+                            arrangement: .horizontal
+                        )
                     )
-                )
+                } else {
+                    ActionButton(
+                        buttonLayout: .dual(
+                            primaryTitle: "완료",
+                            primaryAction: { store.send(.completeButtonTapped) },
+                            secondaryTitle: "다시 투표하기",
+                            secondaryAction: { store.send(.revoteButtonTapped) },
+                            arrangement: .horizontal
+                        )
+                    )
+                }
             }
         }
         .padding(.bottom, Spacing.spacing200)
@@ -317,169 +256,7 @@ private struct PlaceVoteButtonArea: View {
     }
 }
 
-// MARK: - Place Vote Row
-
-private struct PlaceVoteRow: View {
-    let candidate: PlaceVoteCandidate
-    let mode: PlaceVoteParticipationFeature.Mode
-    let isSelected: Bool
-    let isTop: Bool
-    /// row 전체 탭. 해당 장소 좌표로 지도를 포커싱한다.
-    let onTap: () -> Void
-    /// checkbox 탭. 후보를 투표 대상으로 선택한다.
-    let onSelect: () -> Void
-
-    var body: some View {
-        HStack(spacing: Spacing.spacing200) {
-            if mode == .voting {
-                Checkbox(
-                    variant: .circle,
-                    state: isSelected ? .enabled : .disabled,
-                    size: .small
-                )
-                .contentShape(Rectangle())
-                .onTapGesture { onSelect() }
-            }
-
-            PlaceCategoryIcon(category: candidate.categoryLabel)
-                .frame(width: Metric.categoryIconLength, height: Metric.categoryIconLength)
-
-            VStack(alignment: .leading, spacing: Spacing.spacing50) {
-                BangawoText(candidate.name, textStyle: .bodyMediumEmphasized)
-                    .foregroundStyle(Colors.gray800)
-
-                BangawoText(candidate.address, textStyle: .bodySmall)
-                    .foregroundStyle(Colors.gray700)
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-
-            if mode == .voted {
-                voteResult
-            }
-        }
-        .padding(Spacing.spacing250)
-        .background(
-            RoundedRectangle(cornerRadius: BorderRadius.borderRadius250)
-                .fill(backgroundColor)
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: BorderRadius.borderRadius250)
-                .stroke(borderColor, lineWidth: BorderWidth.borderWidth100)
-        )
-        .contentShape(Rectangle())
-        .onTapGesture { onTap() }
-    }
-
-    private var voteResult: some View {
-        VStack(alignment: .trailing, spacing: Spacing.spacing50) {
-            if isTop {
-                Badge("1위", variant: .solid, size: .small)
-            }
-
-            BangawoText("\(candidate.voteCount)명 투표", textStyle: .bodySmall)
-                .foregroundStyle(Colors.gray700)
-        }
-    }
-
-    private var backgroundColor: Color {
-        if mode == .voted, isTop { return Colors.orange100 }
-        if mode == .voting, isSelected { return Colors.gray200 }
-
-        return Colors.gray50
-    }
-
-    private var borderColor: Color {
-        if mode == .voted, isTop { return Colors.orange300 }
-        if mode == .voting, isSelected { return Colors.gray300 }
-
-        return Colors.gray200
-    }
-}
-
-// MARK: - Category Icon
-
-private struct PlaceCategoryIcon: View {
-    let category: PlaceCategory
-
-    var body: some View {
-        category.pinIcon
-            .resizable()
-            .scaledToFit()
-            .clipShape(Circle())
-    }
-}
-
-// MARK: - Deadline Description
-
-private struct DeadlineDescription: View {
-    let deadline: String?
-
-    @State private var now = Date()
-
-    private let timer = Timer.publish(every: 1, on: .main, in: .common).autoconnect()
-
-    var body: some View {
-        Group {
-            if let remaining = PlaceVoteCountdown.remaining(deadline, now: now) {
-                if remaining.isExpired {
-                    BangawoText("투표 종료", textStyle: .bodySmall)
-                        .foregroundStyle(Colors.gray700)
-                } else {
-                    HStack(spacing: 0) {
-                        BangawoText(remaining.timeText, textStyle: .bodySmall)
-                            .foregroundStyle(Colors.orange600)
-
-                        BangawoText(" 후 종료", textStyle: .bodySmall)
-                            .foregroundStyle(Colors.gray700)
-                    }
-                }
-            }
-        }
-        .onReceive(timer) { now = $0 }
-    }
-}
-
-// MARK: - Countdown Formatter
-
-private enum PlaceVoteCountdown {
-    struct Remaining {
-        let timeText: String
-        let isExpired: Bool
-    }
-
-    /// `now` 기준 마감까지 남은 시간 컴포넌트. 파싱 실패 시 nil(표시 안 함).
-    static func remaining(_ raw: String?, now: Date) -> Remaining? {
-        guard
-            let raw,
-            let date = DateFormatterStore.date(
-                from: raw,
-                format: "yyyy-MM-dd'T'HH:mm:ss",
-                locale: "en_US_POSIX",
-                timeZone: "UTC"
-            )
-        else { return nil }
-
-        let total = max(0, Int(date.timeIntervalSince(now)))
-        guard total > 0 else { return Remaining(timeText: "", isExpired: true) }
-
-        let days = total / 86400
-        let hours = (total % 86400) / 3600
-        let minutes = (total % 3600) / 60
-        let seconds = total % 60
-
-        let time = String(format: "%02d:%02d:%02d", hours, minutes, seconds)
-        let timeText = days > 0 ? "\(days)일 \(time)" : time
-
-        return Remaining(timeText: timeText, isExpired: false)
-    }
-}
-
 // MARK: - Constants
-
-private enum Metric {
-    /// 카테고리 아이콘 한 변 길이.
-    static let categoryIconLength: CGFloat = 32
-}
 
 private enum Constant {
     /// 후보 좌표가 없을 때 사용하는 기본 지도 중심(서울 시청).
